requirements.txt
pandas
matplotlib
seaborn
wordcloud
scikit-learn
streamlit
pytesseract
nltk
Pillow
